backend/requirements.txt
# packages in environment at /opt/anaconda3/envs/oil_spill_detector_env:
#
# Name                    Version                   Build  Channel
asgiref                   3.8.1              pyhd8ed1ab_0    conda-forge
bzip2                     1.0.8                h99b78c6_7    conda-forge
ca-certificates           2024.8.30            hf0a4a13_0    conda-forge
django                    5.1.1              pyhd8ed1ab_0    conda-forge
djangorestframework       3.15.2             pyhd8ed1ab_0    conda-forge
libexpat                  2.6.3                hf9b8971_0    conda-forge
libffi                    3.4.2                h3422bc3_5    conda-forge
libsqlite                 3.46.1               hc14010f_0    conda-forge
libzlib                   1.3.1                hfb2fe0b_1    conda-forge
ncurses                   6.5                  h7bae524_1    conda-forge
openssl                   3.3.2                h8359307_0    conda-forge
pip                       24.2               pyh8b19718_1    conda-forge
python                    3.12.5          h30c5eda_0_cpython    conda-forge
readline                  8.2                  h92ec313_1    conda-forge
setuptools                73.0.1             pyhd8ed1ab_0    conda-forge
sqlparse                  0.5.1              pyhd8ed1ab_0    conda-forge
tk                        8.6.13               h5083fa2_1    conda-forge
typing_extensions         4.12.2             pyha770c72_0    conda-forge
tzdata                    2024a                h8827d51_1    conda-forge
wheel                     0.44.0             pyhd8ed1ab_0    conda-forge
xz                        5.2.6                h57fd34a_0    conda-forge
